+++ packages/information-architect/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @auto-engineer/information-architect
 
+## 0.6.4
+
+### Patch Changes
+
+- version test
+- Updated dependencies
+  - @auto-engineer/message-bus@0.5.4
+
+## 0.6.3
+
+### Patch Changes
+
+- fix version report
+- Updated dependencies
+  - @auto-engineer/message-bus@0.5.3
+
 ## 0.6.2
 
 ### Patch Changes